+++ CHANGELOG.md
@@ -141,8 +141,8 @@
   [Martin Redington](https://github.com/mildm8nnered)
   [#4200](https://github.com/realm/SwiftLint/issues/4200)
 
-* Add a new `if_let_shadowing` opt-in rule that triggers in Swift 5.7 when a
-  shadowing optional binding is created in an if- or guard-statement.  
+* Add a new `shorthand_optional_binding` opt-in rule that triggers in Swift 5.7
+  when a shadowing optional binding is created in an `if` or `guard` statement.  
   [SimplyDanny](https://github.com/SimplyDanny)
   [#4202](https://github.com/realm/SwiftLint/issues/4202)